+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 ## 1.48.1 (Unreleased)
+
+FEATURES:
+
+* **New Resource**: `tencentcloud_cos_bucket_policy`
+
+ENHANCEMENTS:
+
+* Resource: `tencentcloud_kubernetes_as_scaling_group` support `max_size` and `min_size` modification.
+
 ## 1.48.0 (November 20, 2020)
 
 FEATURES: